# packages/contracts/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "CoCoLo API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/auth/invitations": {
      "get": {
        "operationId": "listAuthInvitations",
        "summary": "チームの招待一覧を取得",
        "responses": {
          "200": {
            "description": "招待一覧",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      },
      "post": {
        "operationId": "createAuthInvitation",
        "summary": "対象memberへの招待を発行",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "招待を発行",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationCreateResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/auth/invitations/accept": {
      "post": {
        "operationId": "acceptAuthInvitation",
        "summary": "OAuth認証済み利用者が招待を受諾",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvitationAcceptInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "招待を受諾",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationAcceptResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/auth/invitations/{invitationId}/revoke": {
      "post": {
        "operationId": "revokeAuthInvitation",
        "summary": "招待を取り消す",
        "parameters": [
          {
            "name": "invitationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "招待を取り消す",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvitationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "InvitationCreateInput": {
        "type": "object",
        "required": [
          "memberId",
          "role",
          "linkType",
          "relationship"
        ],
        "additionalProperties": false,
        "properties": {
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "const": "guardian"
          },
          "linkType": {
            "type": "string",
            "enum": [
              "self",
              "guardian"
            ]
          },
          "relationship": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "expiresInHours": {
            "type": "integer",
            "minimum": 1,
            "maximum": 168,
            "default": 72
          }
        }
      },
      "InvitationAcceptInput": {
        "type": "object",
        "required": [
          "token",
          "provider"
        ],
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "minLength": 32,
            "maxLength": 256
          },
          "provider": {
            "type": "string",
            "enum": [
              "google",
              "line"
            ]
          }
        }
      },
      "InvitationItem": {
        "type": "object",
        "required": [
          "id",
          "memberId",
          "role",
          "linkType",
          "relationship",
          "status",
          "expiresAt",
          "acceptedAt"
        ],
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "memberId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "const": "guardian"
          },
          "linkType": {
            "type": "string",
            "enum": [
              "self",
              "guardian"
            ]
          },
          "relationship": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "accepted",
              "expired",
              "revoked"
            ]
          },
          "expiresAt": {
            "type": "string",
            "format": "date-time"
          },
          "acceptedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
      "InvitationListResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvitationItem"
            }
          }
        }
      },
      "InvitationResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/InvitationItem"
          }
        }
      },
      "InvitationCreateResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "id",
              "memberId",
              "role",
              "relationship",
              "token",
              "expiresAt"
            ],
            "additionalProperties": false,
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid"
              },
              "memberId": {
                "type": "string",
                "format": "uuid"
              },
              "role": {
                "type": "string",
                "const": "guardian"
              },
              "linkType": {
                "type": "string",
                "enum": [
                  "self",
                  "guardian"
                ]
              },
              "relationship": {
                "type": "string",
                "minLength": 1,
                "maxLength": 100
              },
              "token": {
                "type": "string",
                "minLength": 32,
                "maxLength": 256
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        }
      },
      "InvitationAcceptResponse": {
        "type": "object",
        "required": [
          "data"
        ],
        "additionalProperties": false,
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "tenantId",
              "memberId",
              "role",
              "linkType",
              "linkStatus"
            ],
            "additionalProperties": false,
            "properties": {
              "tenantId": {
                "type": "string",
                "format": "uuid"
              },
              "memberId": {
                "type": "string",
                "format": "uuid"
              },
              "role": {
                "type": "string",
                "const": "guardian"
              },
              "linkType": {
                "type": "string",
                "enum": [
                  "self",
                  "guardian"
                ]
              },
              "linkStatus": {
                "type": "string",
                "const": "active"
              }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "additionalProperties": false,
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "details",
              "requestId"
            ],
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1,
                "maxLength": 128
              },
              "message": {
                "type": "string",
                "minLength": 1,
                "maxLength": 512
              },
              "details": {},
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "入力値が不正です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "認証が必要です。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "この操作を実行する権限がありません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Conflict": {
        "description": "リクエストが現在の状態と競合しました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "指定されたリソースが見つかりません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "リクエスト数の上限を超えました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "内部エラーが発生しました。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "ServiceUnavailable": {
        "description": "外部サービスまたは依存サービスが利用できません。",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    }
  },
...
}